potential fix for time period sanity check
the logic was initially comparing 'count' (time period in seconds) to 'ttl / duration' (max # of stored datapoints for this granularity). This produces a problem when requesting data like ts.getHits('stats', '5minutes', ts.minutes(10) ...). Even with this which resolves to 2 data point you get an error.

We could fix the math and it reduces to just: time_period_requested (in seconds) > ttl_for_granularity . if you request more data than is stored, you get an error.